configurando swagger na rota geracao
