+++ src/server.js
@@ -2,8 +2,8 @@
 const express = require("express");
 const cors = require("cors");
 const morgan = require("morgan");
-const swaggerUi = require('swagger-ui-express');
-const swaggerSpec = require('../swaggerConfig');
+const swaggerUi = require("swagger-ui-express");
+const swaggerSpec = require("../swagger"); 
 const { config } = require("dotenv");
 config();
 
@@ -21,6 +21,7 @@ class Server {
     app.use(cors());
     app.use(express.json());
     app.use(morgan("dev"))
+    
     app.use('/api-docs', swaggerUi.serve, swaggerUi.setup(swaggerSpec));
 
   }
